feat: add configurable GitHub Pages support
While the default behaviour stays the same as before (the built dist is uploaded as an artifact of the workflow run), another workflow option allows to directly deploy to GitHub Pages, all the user has to do is check an additional checkbox when launching the workflow manually.

The base URL is taken into account and is automatically determined to be the name of the repository. It can be disabled (set to empty) with another new workflow parameter.